pyproject.toml
[project]
name = "elt-pipeline"
version = "0.1.0"
description = "Production-grade ELT pipeline: Real-time Crypto Prices → Polars → DuckDB → Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars>=0.20.0",
    "duckdb>=0.10.0",
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
